pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vpn-leaks"
version = "0.1.0"
description = "VPN leak benchmarking harness (client-side observable privacy)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "playwright>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4.0",
]

[project.scripts]
vpn-leaks = "vpn_leaks.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["vpn_leaks"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
